+++ .packages
@@ -3,29 +3,29 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2021-11-20 08:25:41.846836.
-async:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.1/lib/
-boolean_selector:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
-characters:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.1.0/lib/
-charcode:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/lib/
-clock:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/lib/
-collection:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.15.0/lib/
-cupertino_icons:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.4/lib/
-fake_async:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.2.0/lib/
-flutter:file:///home/nhanaqwahme/development/flutter/packages/flutter/lib/
-flutter_lints:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
-flutter_test:file:///home/nhanaqwahme/development/flutter/packages/flutter_test/lib/
-lints:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/lib/
-matcher:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.10/lib/
-meta:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/lib/
-path:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.0/lib/
-sky_engine:file:///home/nhanaqwahme/development/flutter/bin/cache/pkg/sky_engine/lib/
-source_span:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.1/lib/
-stack_trace:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
-stream_channel:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
-string_scanner:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/lib/
-term_glyph:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/lib/
-test_api:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.2/lib/
-typed_data:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/typed_data-1.3.0/lib/
-vector_math:file:///home/nhanaqwahme/development/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.0/lib/
+# Generated by pub on 2021-11-15 10:07:57.487531.
+async:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.1/lib/
+boolean_selector:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
+characters:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.1.0/lib/
+charcode:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/charcode-1.3.1/lib/
+clock:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/clock-1.1.0/lib/
+collection:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/collection-1.15.0/lib/
+cupertino_icons:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/cupertino_icons-1.0.4/lib/
+fake_async:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/fake_async-1.2.0/lib/
+flutter:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/packages/flutter/lib/
+flutter_lints:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/flutter_lints-1.0.4/lib/
+flutter_test:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/packages/flutter_test/lib/
+lints:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/lints-1.0.1/lib/
+matcher:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/matcher-0.12.10/lib/
+meta:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/meta-1.7.0/lib/
+path:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/path-1.8.0/lib/
+sky_engine:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/bin/cache/pkg/sky_engine/lib/
+source_span:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/source_span-1.8.1/lib/
+stack_trace:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/stack_trace-1.10.0/lib/
+stream_channel:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/stream_channel-2.1.0/lib/
+string_scanner:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/string_scanner-1.1.0/lib/
+term_glyph:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/term_glyph-1.2.0/lib/
+test_api:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/test_api-0.4.2/lib/
+typed_data:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/typed_data-1.3.0/lib/
+vector_math:file:///C:/Flutter/flutter_windows_2.5.3-stable/flutter/.pub-cache/hosted/pub.dartlang.org/vector_math-2.1.0/lib/
 login_ui:lib/
