--- android/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Sat Nov 20 08:26:11 GMT 2021
-gradle.version=6.7
