+++ android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=/home/nhanaqwahme/Android/Sdk
-flutter.sdk=/home/nhanaqwahme/development/flutter
+sdk.dir=C:\\Users\\USER\\AppData\\Local\\Android\\Sdk
+flutter.sdk=C:\\Flutter\\flutter_windows_2.5.3-stable\\flutter
 flutter.buildMode=debug
 flutter.versionName=1.0.0
 flutter.versionCode=1
