+++ lib/main.dart
@@ -1,6 +1,212 @@
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:login_ui/layOutScreen/loginscreen.dart';
 
 void main() {
   runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Login UI',
+      debugShowCheckedModeBanner: false,
+      home: LocationApp(),
+    );
+  }
+}
+
+class LocationApp extends StatefulWidget {
+  @override
+  _LocationAppState createState() => _LocationAppState();
+}
+
+class _LocationAppState extends State<LocationApp> {
+  get child => null;
+
+  get _color => null;
+
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SingleChildScrollView(
+        child: Container(
+          constraints: BoxConstraints(
+            maxHeight: MediaQuery.of(context).size.height,
+            maxWidth: MediaQuery.of(context).size.width,
+          ),
+          decoration: BoxDecoration(
+              gradient: LinearGradient(
+            colors: [
+              Colors.green.shade500,
+              Colors.red.shade500,
+           ],
+           begin: Alignment.topLeft,
+           end: Alignment.centerRight,
+          ),
+        ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Expanded(
+              flex: 2,
+              child: Padding(
+                padding: const EdgeInsets.symmetric(
+                  vertical: 36.0, horizontal: 24.0),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    Container(
+                      margin: const EdgeInsets.all(8),
+                      width: 100.0,
+                      height: 100.0,
+                      decoration: BoxDecoration(
+                        color: this._color,
+                        borderRadius: BorderRadius.circular(10),
+                      ),
+                      child: const FlutterLogo(),
+                    ),
+                    SizedBox(
+                      height: 10.0,
+                    ),
+                     Text(
+                     "Group 7 Capstone Project",
+                      style: TextStyle(
+                        color: Colors.white,
+                        fontSize: 20.0,
+                        fontWeight: FontWeight.w300
+                      ),
+                    ),
+                 ],
+                ),
+              ),
+            ),
+            Expanded(
+              flex: 5,
+              child:Container(
+                width: double.infinity,
+                decoration: BoxDecoration(
+                    color: Colors.white,
+                    borderRadius: BorderRadius.only(
+                     topLeft: Radius.circular(20),
+                     topRight: Radius.circular(20),
+                    )),
+                child: Padding(
+                  padding: const EdgeInsets.all(24.0),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      TextField(
+                        keyboardType: TextInputType.emailAddress,
+                        decoration: InputDecoration(
+                          border: OutlineInputBorder(
+                            borderRadius: BorderRadius.circular(8.0),
+                            borderSide: BorderSide.none,
+                          ),
+                          filled: true,
+                          fillColor: Color(0xFFe7edeb),
+                          hintText: "E-mail",
+                          prefixIcon: Icon(
+                            Icons.email,
+                            color: Colors.grey[600],
+                          )
+                        ),
+                      ),
+                      SizedBox(
+                        height: 20.0,
+                      ),
+                      TextField(
+                        keyboardType: TextInputType.emailAddress,
+                        decoration: InputDecoration(
+                          border: OutlineInputBorder(
+                            borderRadius: BorderRadius.circular(8.0),
+                            borderSide: BorderSide.none,
+                          ),
+                          filled: true,
+                          fillColor: Color(0xFFe7edeb),
+                          hintText: "Password",
+                          prefixIcon: Icon(
+                            Icons.lock,
+                            color: Colors.grey[600],
+                          )
+                        ),
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.end,
+                        crossAxisAlignment: CrossAxisAlignment.end,
+                        children: [
+                          FlatButton(
+                          child: Text(
+                            "Forget your password?", 
+                              style: TextStyle(
+                              color: Colors.green[800],
+                              decoration: TextDecoration.underline,
+                              ),),
+                            onPressed: () {
+                          },
+                         )     
+                        ],
+                      ),
+                      SizedBox(
+                        height: 50.0,
+                      ),
+                      Container(
+                        width: double.infinity,
+                        child: RaisedButton(
+                          onPressed: () {},
+                          color: Colors.green[600],
+                          child: Padding(
+                            padding: 
+                                const EdgeInsets.symmetric(vertical: 16.0),
+                            child: Text(
+                              "Login",
+                              style: TextStyle(
+                                color: Colors.white,
+                                fontSize: 16.0,
+                              )
+                            ),
+                          ),
+                        ),
+                      ),
+                      SizedBox(
+                        height: 10.0,
+                      ),
+                       Row (
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: [
+                            Text(
+                              "Don't have an account?"),
+                            FlatButton(
+                              child: Text("Sign up",
+                                style: TextStyle(
+                                  color: Colors.red[800],
+                                  fontSize: 15.0,
+                                  fontWeight: FontWeight.w600,
+                                ),
+                              ),
+                              onPressed: () {
+                              },
+                            ),
+                          ],
+                        ),
+                      
+                    ],
+                  ),
+                ),
+              ),
+            )
+          ],
+          ),
+        ),
+      ),
+    );
+  }
 }
